Fix parsing long values

+++ ppm-war/src/main/java/de/blogspot/soahowto/ppm/ListSelectionUtils.java
@@ -43,7 +43,7 @@ public class ListSelectionUtils {
         List<String> filteredList = new ArrayList<>();
 
         for (String id: list) {
-            if (Integer.parseInt(id) % modulus == remainder) {
+            if (Long.parseLong(id) % modulus == remainder) {
                 filteredList.add(id);
             }
         }
